+++ src/main/java/com/reminderx/accountservice/service/FundService.java
@@ -1,8 +1,11 @@
 package com.reminderx.accountservice.service;
 
 public interface FundService {
-    double balance(String number);
-    double deposit(String number, double fund);
-    double withdraw(String number, double fund);
-    void transfer(String from, String to, double fund);
+    long balance(String number);
+
+    long deposit(String number, long fund);
+
+    long withdraw(String number, long fund);
+
+    void transfer(String from, String to, long fund);
 }
